+++ CHANGELOG.md
@@ -8,3 +8,5 @@ This mod is distributed under Version 2 of the GPL.  A copy of the GPL is includ
 
 - Add Gyromatic Icemaker in Gnomeregan
 - Replace "The Darksoul" in "Misc Sets" with "Shard of the Gods"
+- Add General Rajaxx encounter loot
+- Add Warblade of the Hakkari (Off hand) for Hakkar